requirements.txt
colorama==0.4.6
mido==1.3.3
numpy==2.2.6
packaging==25.0
pyfluidsynth==1.3.4
python-rtmidi==1.5.8
